+++ crates/adapters/okx/src/common/enums.rs
@@ -1315,6 +1315,20 @@ pub fn conditional_order_to_algo_type(order_type: OrderType) -> anyhow::Result<O
     Deserialize,
 )]
 #[serde(rename_all = "snake_case")]
+#[cfg_attr(
+    feature = "python",
+    pyo3::pyclass(
+        eq,
+        eq_int,
+        module = "nautilus_trader.core.nautilus_pyo3.okx",
+        from_py_object,
+        rename_all = "SCREAMING_SNAKE_CASE",
+    )
+)]
+#[cfg_attr(
+    feature = "python",
+    pyo3_stub_gen::derive::gen_stub_pyclass_enum(module = "nautilus_trader.adapters.okx")
+)]
 pub enum OKXAlgoOrderStatus {
     Live,
     Pause,